Use primitive int/boolean where possible in the Java client.
R=brianwilkerson@google.com, jwren@google.com
BUG=

Review URL: https://codereview.chromium.org//538253002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@39891 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/editor/tools/plugins/com.google.dart.tools.core_test/src/com/google/dart/tools/core/completion/DartCompletionReceiverTest.java
@@ -39,10 +39,9 @@ public class DartCompletionReceiverTest extends TestCase {
     }
 
     @Override
-    public void completion_getSuggestions(String file, Integer offset,
-        GetSuggestionsConsumer consumer) {
+    public void completion_getSuggestions(String file, int offset, GetSuggestionsConsumer consumer) {
       assertEquals(testFile, file);
-      assertEquals(testOffset, offset.intValue());
+      assertEquals(testOffset, offset);
       this.consumer = consumer;
       // Send results asynchronously
       Thread thread = new Thread(getClass().getSimpleName() + " suggestion results") {
